+++ CHANGELOG.md
@@ -118,6 +118,7 @@
 98. Added SHA512 Check to Installer Downloads.
 99. Fixed Issue when using the Installer to install v0.2.
 100. Fixed Issue with Releases Selection in Installer not being properly Sorted.
+101. Added Shortcut Support to Installer.
 
 ---